# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CDT Jamaica Management API with Odoo Community integration
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /tuition/fees:
    get:
      operationId: listTuitionFees
      tags: [tuition]
      summary: List student tuition and fee balances from Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: ["string", "null"]
            enum: [draft, due, partial, paid, cancelled, null]
      responses:
        "200":
          description: Student tuition fee ledger
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TuitionFee"
        "503":
          description: Odoo integration is not configured
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # ── Performances ───────────────────────────────────────────────────────────
...
  /tasks:
    get:
      operationId: listTasks
      tags: [tasks]
      summary: List tasks from Odoo
      parameters:
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: stage
          schema:
            type: ["string", "null"]
            enum: [todo, in_progress, done, cancelled, null]
      responses:
        "200":
          description: List of tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Task"
    post:
      operationId: createTask
      tags: [tasks]
      summary: Create a task in Odoo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskInput"
      responses:
        "201":
          description: Created task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
  /tasks/{id}:
    patch:
      operationId: updateTask
      tags: [tasks]
      summary: Update a task in Odoo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaskUpdate"
      responses:
        "200":
          description: Updated task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    delete:
      operationId: deleteTask
      tags: [tasks]
      summary: Delete a task in Odoo
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted

  # ── Reports ────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TuitionFee:
      type: object
      properties:
        id:
          type: integer
        reference:
          type: string
        studentId:
          type: integer
        studentName:
          type: string
        feeSchedule:
          type: ["string", "null"]
        feeStructure:
          type: string
        dueDate:
          type: string
          format: date
        amount:
          type: number
        amountPaid:
          type: number
        balance:
          type: number
        currency:
          type: string
        status:
          type: string
          enum: [draft, due, partial, paid, cancelled]
        overdue:
          type: boolean
        invoiceId:
          type: ["integer", "null"]
        invoiceName:
          type: ["string", "null"]
        invoiceState:
          type: ["string", "null"]
          enum: [draft, posted, cancelled, null]
        paymentState:
          type: ["string", "null"]
          enum: [not_paid, in_payment, paid, partial, reversed, null]
        contactId:
          type: ["integer", "null"]
        programName:
          type: ["string", "null"]
        classIds:
          type: array
          items:
            type: integer
        classNames:
          type: array
          items:
            type: string
        sampleData:
          type: boolean
      required:
        - id
        - reference
        - studentId
        - studentName
        - feeSchedule
        - feeStructure
        - dueDate
        - amount
        - amountPaid
        - balance
        - currency
        - status
        - overdue
        - invoiceId
        - invoiceName
        - invoiceState
        - paymentState
        - contactId
        - programName
        - classIds
        - classNames
        - sampleData

    # ── Invoice ────────────────────────────────────────────────────────────
...
    Task:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        stage:
          type: string
          enum: [todo, in_progress, done, cancelled]
        assigneeName:
          type: ["string", "null"]
        deadline:
          type: ["string", "null"]
        priority:
          type: string
          enum: [low, normal, high]
        projectName:
          type: ["string", "null"]
        odooId:
          type: ["string", "null"]
        createdAt:
          type: string
      required: [id, title, stage, priority, createdAt]
    TaskInput:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        stage:
          type: string
          enum: [todo, in_progress, done, cancelled]
        assigneeName:
          type: string
        deadline:
          type: string
        priority:
          type: string
          enum: [low, normal, high]
        projectName:
          type: string
      required: [title]
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        stage:
          type: string
          enum: [todo, in_progress, done, cancelled]
        assigneeName:
          type: string
        deadline:
          type: string
        priority:
          type: string
          enum: [low, normal, high]
        projectName:
          type: string

    # ── Reports ────────────────────────────────────────────────────────────
...